✨ refactor: refactor the whole RAG api
