+++ app/constants.py
@@ -1,14 +1,11 @@
 from enum import Enum
 
-
 class MESSAGES(str, Enum):
     DEFAULT = lambda msg="": f"{msg if msg else ''}"
-
 
 class ERROR_MESSAGES(str, Enum):
     def __str__(self) -> str:
         return super().__str__()
-
     DEFAULT = lambda err="": f"Something went wrong :/\n{err if err else ''}"
     PANDOC_NOT_INSTALLED = "Pandoc is not installed on the server. Please contact your administrator for assistance."
     OPENAI_NOT_FOUND = lambda name="": f"OpenAI API was not found"
