+++ app/db/mongo.py
@@ -1,10 +1,6 @@
-import logging
 from pymongo import MongoClient
 from pymongo.errors import PyMongoError
-from config import ATLAS_MONGO_DB_URI
-
-logger = logging.getLogger(__name__)
-
+from app.config import ATLAS_MONGO_DB_URI, logger
 
 async def mongo_health_check() -> bool:
     try:
