+++ app/models.py
@@ -1,22 +1,18 @@
 import hashlib
 from enum import Enum
-from pydantic import BaseModel
 from typing import Optional, List
-
+from pydantic import BaseModel
 
 class DocumentResponse(BaseModel):
     page_content: str
     metadata: dict
-
 
 class DocumentModel(BaseModel):
     page_content: str
     metadata: Optional[dict] = {}
 
     def generate_digest(self):
-        hash_obj = hashlib.md5(self.page_content.encode())
-        return hash_obj.hexdigest()
-
+        return hashlib.md5(self.page_content.encode()).hexdigest()
 
 class StoreDocument(BaseModel):
     filepath: str
@@ -24,20 +20,25 @@ class StoreDocument(BaseModel):
     file_content_type: str
     file_id: str
 
-
 class QueryRequestBody(BaseModel):
     query: str
     file_id: str
     k: int = 4
     entity_id: Optional[str] = None
 
-
 class CleanupMethod(str, Enum):
     incremental = "incremental"
     full = "full"
-
 
 class QueryMultipleBody(BaseModel):
     query: str
     file_ids: List[str]
     k: int = 4
+
+class EmbeddingsProvider(str, Enum):
+    OPENAI = "openai"
+    AZURE = "azure"
+    HUGGINGFACE = "huggingface"
+    HUGGINGFACETEI = "huggingfacetei"
+    OLLAMA = "ollama"
+    BEDROCK = "bedrock"
