+++ app/rag_utils.py
@@ -1,0 +1,1 @@
+from langchain_core.runnables.config import run_in_executor
